map to club_mf_device add, change ruby to 2.2.1 and rails to 4.2.3

+++ db/migrate/20150629193323_add_latitude_and_longitude_to_club.rb
@@ -1,6 +1,6 @@
 class AddLatitudeAndLongitudeToClub < ActiveRecord::Migration
   def change
-    add_column :clubs, :lat, :decimal
-    add_column :clubs, :lng, :decimal
+    add_column :clubs, :lat, :float
+    add_column :clubs, :lng, :float
   end
 end
