+++ db/migrate/20150630212654_add_latitude_and_longitude_to_club_mf_device.rb
@@ -1,6 +1,6 @@
 class AddLatitudeAndLongitudeToClubMfDevice < ActiveRecord::Migration
   def change
-    add_column :club_mf_devices, :lat, :decimal
-    add_column :club_mf_devices, :lng, :decimal
+    add_column :club_mf_devices, :lat, :float
+    add_column :club_mf_devices, :lng, :float
   end
 end
